Fix imprecise database column types
* Don't allow `planet_has_cargo.amount` or `ship_has_cargo.amount` to
  be null, since these rows should only exist if there is some amount
  of cargo. Also change them from int -> smallint because we don't need
  to support such large values. Fixes the following warnings:

  > Smr\DatabaseRecord::getInt returns type int, but field `amount` has type int<0, 4294967295>|null
  > Smr\DatabaseRecord::getInt returns type int, but field `amount` has type int<0, 4294967295>|null

* Change `weighted_random.weighting` from int -> float, because we use
  it as a float. Fixes the following warning:

  > Query error: Column "weighted_random.weighting" expects value type int, got type float

+++ src/lib/Smr/WeightedRandom.php
@@ -60,7 +60,7 @@ class WeightedRandom {
 			'type_id' => $db->escapeNumber($typeID),
 		]);
 		if ($dbResult->hasRecord()) {
-			$this->weighting = $dbResult->record()->getInt('weighting');
+			$this->weighting = $dbResult->record()->getFloat('weighting');
 		} else {
 			$this->weighting = 0;
 		}
